telegram: Do not notify when a proxyed message gets sent
The main benefit here is that instead of double notifications for others
and an unhelpful notification for the sender, there should only be a
single notification sent when the original unproxied message is sent.

Furthermore, if the original message was set not to notify, the proxied
message will now not undo that choice by notifying regardless.