discord: implement force proxied emoji (resolves #11)
This adds a new feature to Séance where reference user reactions of
emoji can automatically be proxied. Any emoji provided by the
`proxied-emoji` config value will be proxied.

This removes the ability of the reference user to react with those emoji
and is primarily assumed to be used for things like custom heart
reactions that are unique per-user, making their usage easier.

The valid config value is a comma or whitespace separated list of
unicode emoji and Discord custom emoji IDs that should be handled this
way. `*` may also be used to indicate that *all* reactions by the
reference user should be proxied.

This commit also revises the README to more clearly list the available
configuration options.