discord: add `!reproxy` command (resolves #20)
This adds support for a `!reproxy` command which makes it easy to
"reproxy" messages from a set of provided users and proxy it from
ourselves.

This feature is dangerous, and requires configuring a
`valid-reproxy-targets` list to specify user IDs that can be authors of
the message targeted to be reproxied. The reference user is always on
this list however.

Because of the danger for collision, if a command prefix is set it is
required to be used with the command to avoid multiple Séance instances
competing for a reproxy.